textaura-api/requirements.txt
fastapi
python-dotenv
uvicorn
youtube-transcript-api
google-genai
